+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Change Log
 
-### 3.19.1(59)
+### [3.19.1(59)](https://github.com/VakhoKontridze/VCore/releases/tag/3.19.1) — *2022 08 11*
 
 Extensions